CORREGUIR A DOUBLE TARIFA PARTICIPANTES

+++ look.domain/entities/admin/ProyectoParticipante.cs
@@ -14,7 +14,7 @@ namespace look.domain.entities.admin
         
         public int? CarId { get; set; }
         
-        public int? PerTarifa { get; set; }
+        public double? PerTarifa { get; set; }
         
         public int PrfId {get; set; }
         public int? TarifarioId { get; set; }
